+++ marketing-analytics/activation/gmp-googleads-connector/src/api_handlers/ga_data_import.js
@@ -21,7 +21,7 @@
 
 const {File} = require('@google-cloud/storage');
 const {
-  api: {analytics: {Analytics, DataImportConfig}},
+  api: {analytics: {Analytics, DataImportConfig, DataImportClearConfig,}},
   storage: {StorageFile},
 } = require('@google-cloud/nodejs-common');
 
@@ -40,10 +40,13 @@ exports.defaultOnGcs = true;
  * config item dataImportHeader may apply. The function 'sendData' here will
  * automatically attach this line at the beginning of data file before
  * uploading.
+ * 'clearOption' is the configuration to manage how to clear previous uploaded
+ * data files.
  *
  * @typedef {{
  *   dataImportHeader:(string|undefined),
  *   gaConfig:!DataImportConfig,
+ *   clearOption:(!DataImportClearConfig|undefined),
  * }}
  */
 let GoogleAnalyticsConfig;
@@ -61,21 +64,23 @@ exports.GoogleAnalyticsConfig = GoogleAnalyticsConfig;
  *     file.
  * @return {!Promise<!File>} Cloud Storage File to be uploaded.
  */
-const prepareFile = (bucket, fileName, dataImportHeader = undefined) => {
+const prepareFile = async (bucket, fileName, dataImportHeader = undefined) => {
   const storageFile = new StorageFile(bucket, fileName);
   if (!dataImportHeader) {
     console.log(`No head line need to take care for: `, fileName);
-    return Promise.resolve(storageFile.getFile());
+    return storageFile.getFile();
   }
   console.log(`Appends ${dataImportHeader} to the head of: `, fileName);
-  return storageFile.addHeader(dataImportHeader).then((newFileName) => {
-    return new StorageFile(bucket, newFileName).getFile();
-  });
+  const newFileName = await storageFile.addHeader(dataImportHeader);
+  return new StorageFile(bucket, newFileName).getFile();
 };
 
 /**
  * Sends the data or file from a Pub/sub message to Google Analytics Data
- * Import.
+ * Import. It will delete previous uploaded data files if there is a setting
+ * in the 'GoogleAnalyticsConfig'.
+ * @see https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/uploads/
+ *
  * @param {string} message Message data from Pubsub. It could be the
  *     information of the file to be sent out, or a piece of data that need to
  *     be send out.
@@ -84,26 +89,46 @@ const prepareFile = (bucket, fileName, dataImportHeader = undefined) => {
  * @return {!Promise<boolean>} Whether 'records' have been sent out without any
  *     errors.
  */
-exports.sendData = (message, messageId, config) => {
-  let data;
+const sendData = (message, messageId, config) => {
+  const analytics = new Analytics();
+  return sendDataInternal(analytics, message, messageId, config);
+}
+exports.sendData = sendData;
+
+/**
+ * Internal sendData function for test.
+ * @param {!Analytics} analytics Injected Analytics instance.
+ * @param {string} message Message data from Pubsub. It could be the
+ *     information of the file to be sent out, or a piece of data that need to
+ *     be send out.
+ * @param {string} messageId Pub/sub message ID for log.
+ * @param {!GoogleAnalyticsConfig} config
+ * @return {!Promise<boolean>} Whether 'records' have been sent out without any
+ *     errors.
+ */
+const sendDataInternal = async (analytics, message, messageId, config) => {
+  let uploadData = '';
+  const {dataImportHeader} = config;
   try {
-    data = JSON.parse(message);
+    const {bucket, file} = JSON.parse(message);
+    if (bucket) {  // Data is a GCS file.
+      const file = await prepareFile(bucket, file, dataImportHeader);
+      uploadData = file.createReadStream();
+    } else {
+      console.error('Could find bucket in message', message);
+      return false;
+    }
   } catch (error) {
     console.log(`This is not a JSON string. GA Data Import's data not on GCS.`);
-    data = message;
-  }
-  const analytics = new Analytics();
-  let promise;
-  if (data.bucket) {  // Data is a GCS file.
-    promise = prepareFile(data.bucket, data.file, config.dataImportHeader)
-        .then((file) => file.createReadStream());
-  } else {  // Data comes from the message data.
-    if (config.dataImportHeader) {
-      data = config.dataImportHeader + '\n' + data;
+    if (dataImportHeader) {
+      uploadData = dataImportHeader + '\n';
     }
-    promise = Promise.resolve(data);
+    uploadData += message;
   }
-  // noinspection JSCheckFunctionSignatures
-  return promise.then(
-      (data) => analytics.uploadData(data, config.gaConfig, messageId));
+  if (config.clearOption) {
+    await analytics.deleteUploadedData(config.gaConfig, config.clearOption);
+  }
+  return analytics.uploadData(uploadData, config.gaConfig, messageId);
 };
+
+exports.sendDataInternal = sendDataInternal;