+++ marketing-analytics/activation/common-libs/nodejs-common/RELEASE_NOTES.md
@@ -1,16 +1,29 @@
 Release Notes
 ===
 
+0.7.6 (2021-07-24)
+---
+
+### Google Analytic Data Import supports to clear previously uploaded data files.
+
+0.7.2 (2021-06-09)
+---
+
+### Support Non-ASCII data files
+
 0.7.0 (2021-03-03)
 ---
+
 ### Upgrade Google Ads API to v7.0
 
 0.5.1 (2021-04-07)
 ---
+
 ### Add Api for Measurement Protocol Google Analytics 4
 
 0.5.0 (2021-03-03)
 ---
+
 ### Add component of Google Cloud AutoMl Tables API